+++ spec/open_meteo/entities/location_spec.rb
@@ -6,16 +6,16 @@ RSpec.describe OpenMeteo::Entities::Location do
   describe "#to_get_params" do
     subject { location.to_get_params }
 
-    let(:latitude) { 1.1 }
-    let(:longitude) { 2.2 }
+    let(:latitude) { 1.1.to_d }
+    let(:longitude) { 2.2.to_d }
 
     it { is_expected.to eq({ latitude: 1.1, longitude: 2.2 }) }
   end
 
   describe "#validate!" do
     context "when the latitude and longitude are valid" do
-      let(:latitude) { 1.1 }
-      let(:longitude) { 2.2 }
+      let(:latitude) { 1.1.to_d }
+      let(:longitude) { 2.2.to_d }
 
       it "does not raise an error" do
         expect { location.validate! }.not_to raise_error
@@ -23,38 +23,46 @@ RSpec.describe OpenMeteo::Entities::Location do
     end
 
     context "when the latitude is higher than 90" do
-      let(:latitude) { 91 }
-      let(:longitude) { 2.2 }
+      let(:latitude) { 91.to_d }
+      let(:longitude) { 2.2.to_d }
 
       it "raises an error" do
-        expect { location.validate! }.to raise_error(Dry::Struct::Error)
+        expect { location.validate! }.to raise_error(
+          OpenMeteo::Entities::Contracts::ApplicationContract::ValidationError,
+        )
       end
     end
 
     context "when the latitude is lower than -90" do
-      let(:latitude) { -91 }
-      let(:longitude) { 2.2 }
+      let(:latitude) { -91.to_d }
+      let(:longitude) { 2.2.to_d }
 
       it "raises an error" do
-        expect { location.validate! }.to raise_error(Dry::Struct::Error)
+        expect { location.validate! }.to raise_error(
+          OpenMeteo::Entities::Contracts::ApplicationContract::ValidationError,
+        )
       end
     end
 
     context "when the longitude is higher than 180" do
-      let(:latitude) { 1.1 }
-      let(:longitude) { 181 }
+      let(:latitude) { 1.1.to_d }
+      let(:longitude) { 181.to_d }
 
       it "raises an error" do
-        expect { location.validate! }.to raise_error(Dry::Struct::Error)
+        expect { location.validate! }.to raise_error(
+          OpenMeteo::Entities::Contracts::ApplicationContract::ValidationError,
+        )
       end
     end
 
     context "when the longitude is lower than -180" do
-      let(:latitude) { 1.1 }
-      let(:longitude) { -181 }
+      let(:latitude) { 1.1.to_d }
+      let(:longitude) { -181.to_d }
 
       it "raises an error" do
-        expect { location.validate! }.to raise_error(Dry::Struct::Error)
+        expect { location.validate! }.to raise_error(
+          OpenMeteo::Entities::Contracts::ApplicationContract::ValidationError,
+        )
       end
     end
   end