+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Don't validate variables, see https://github.com/open-meteo/open-meteo-website/issues/40
 - Return data in the shape of understandable and manageable entities
+- Change Location latitude/longitude types to coercible decimals
 
 ## [0.0.3] - 2023-11-03